gpio-usbdetect: Add support for ID detection
The existing driver does not support for role
switch based on ID. Add this support to switch to
host mode depending upon the value of the GPIO
qcom,id-det-gpio.
Also the final state table if this GPIO is present
	USB STATE	ID	VBUS
	PERIPHERAL	HIGH	HIGH
	NONE		HIGH	LOW
	HOST		LOW	LOW
	HOST		LOW	HIGH
Also add support for DPDM routing to different ports
based on a switch driven by a GPIO. In host mode
it is routed to one and in peripheral to the other.

Change-Id: I3a6aa36e06ea433318262d26f940b106d79e48d7
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>